+++ server/toilet/src/main/java/learn/toilet/controller/RestroomController.java
@@ -2,9 +2,15 @@ package learn.toilet.controller;
 
 import learn.toilet.domain.RestroomService;
 import learn.toilet.domain.Result;
+import learn.toilet.models.AppUser;
 import learn.toilet.models.Restroom;
+import learn.toilet.security.AppUserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,9 +20,11 @@ import java.util.List;
 @RequestMapping("/api/restroom")
 public class RestroomController {
 
+    private final AppUserService appUserService;
     private final RestroomService restroomService;
 
-    public RestroomController(RestroomService restroomService) {
+    public RestroomController(AppUserService appUserService, RestroomService restroomService) {
+        this.appUserService = appUserService;
         this.restroomService = restroomService;
     }
 
@@ -41,6 +49,20 @@ public class RestroomController {
 
     @PostMapping
     public ResponseEntity<Object> add(@RequestBody Restroom restroom) {
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        // Ensure that the restroom's userId matches the authenticated user's userId
+        if (restroom.getUserId() != authenticatedUserId) {
+            return new ResponseEntity<>("You are not authorized to create a restroom entry for another user.", HttpStatus.FORBIDDEN);
+        }
+
         Result<Restroom> result = restroomService.add(restroom);
         if (result.isSuccess()) {
             return new ResponseEntity<>(result.getPayload(), HttpStatus.CREATED);
@@ -54,6 +76,20 @@ public class RestroomController {
             return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
 
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        // Ensure that the restroom's userId matches the authenticated user's userId
+        if (restroom.getUserId() != authenticatedUserId) {
+            return new ResponseEntity<>("You are not authorized to create a restroom entry for another user.", HttpStatus.FORBIDDEN);
+        }
+
         Result<Restroom> result = restroomService.update(restroom);
         if (result.isSuccess()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -62,7 +98,24 @@ public class RestroomController {
     }
 
     @DeleteMapping("/{restroomId}")
-    public ResponseEntity<Void> deleteById(@PathVariable int restroomId) {
+    public ResponseEntity<Object> deleteById(@PathVariable int restroomId) {
+
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        Restroom restroom = restroomService.findById(restroomId);
+
+        // Ensure that the restroom's userId matches the authenticated user's userId
+        if (restroom.getUserId() != authenticatedUserId) {
+            return new ResponseEntity<>("You are not authorized to create a restroom entry for another user.", HttpStatus.FORBIDDEN);
+        }
+
         Result<Restroom> result = restroomService.deleteById(restroomId);
         if (result.isSuccess()) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
