+++ server/toilet/src/main/java/learn/toilet/controller/ReviewController.java
@@ -3,9 +3,13 @@ package learn.toilet.controller;
 
 import learn.toilet.domain.Result;
 import learn.toilet.domain.ReviewService;
+import learn.toilet.models.AppUser;
 import learn.toilet.models.Review;
+import learn.toilet.security.AppUserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,15 +19,26 @@ import java.util.List;
 @RequestMapping("/api/review")
 public class ReviewController {
 
+    private final AppUserService appUserService;
     private final ReviewService service;
 
-    public ReviewController(ReviewService service) {
+    public ReviewController(AppUserService appUserService, ReviewService service) {
+        this.appUserService = appUserService;
         this.service = service;
     }
 
     @GetMapping("/current")
-    public List<Review> findByUserId(@PathVariable int userId) {
-        return service.findByUserId(userId);
+    public List<Review> findByUserId() {
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        return service.findByUserId(authenticatedUserId);
     }
 
     @GetMapping("/{restroomId}")
@@ -33,6 +48,19 @@ public class ReviewController {
 
     @PostMapping
     public ResponseEntity<Object> add(@RequestBody Review review) {
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        // Ensure that the restroom's userId matches the authenticated user's userId
+        if (review.getUserId() != authenticatedUserId) {
+            return new ResponseEntity<>("You are not authorized to create a review entry for another user.", HttpStatus.FORBIDDEN);
+        }
         Result<Review> result = service.add(review);
         if(result.isSuccess()) {
             return new ResponseEntity<>(result.getPayload(), HttpStatus.CREATED);
@@ -43,6 +71,19 @@ public class ReviewController {
 
     @PutMapping("/{reviewId}")
     public ResponseEntity<Object> update(@PathVariable int reviewId, @RequestBody Review review) {
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        // Ensure that the restroom's userId matches the authenticated user's userId
+        if (review.getUserId() != authenticatedUserId) {
+            return new ResponseEntity<>("You are not authorized to create a restroom entry for another user.", HttpStatus.FORBIDDEN);
+        }
         if(reviewId != review.getReviewId()) {
             return new ResponseEntity<>(HttpStatus.CONFLICT);
         }
@@ -57,6 +98,21 @@ public class ReviewController {
 
     @DeleteMapping("/{reviewId}")
     public ResponseEntity<Object> deleteById(@PathVariable int reviewId) {
+        // Get the authenticated username from the JWT token
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = (String) authentication.getPrincipal();
+        AppUser user = (AppUser) appUserService.loadUserByUsername(username);
+
+        // Get user id from AppUser
+        int authenticatedUserId = user.getAppUserId();
+
+        Review review = service.findById(authenticatedUserId);
+
+        // Ensure that the restroom's userId matches the authenticated user's userId
+        if (review.getUserId() != authenticatedUserId) {
+            return new ResponseEntity<>("You are not authorized to create a restroom entry for another user.", HttpStatus.FORBIDDEN);
+        }
         if(service.deleteById(reviewId)) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
